💫 UPDATE: refactors MSD secondary menu
