+++ client/dashboard/app/secondary-menu/notifications-dropdown/style.scss
